requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.1
click>=8.1.7
tabulate>=0.9.0
pandas>=2.1.4
matplotlib>=3.8.2
SQLAlchemy>=2.0.25
alembic>=1.13.1
pytest>=7.4.4
pytest-cov>=4.1.0
black>=23.12.1
isort>=5.13.2
flake8>=7.0.0
pywin32>=306
